refactor: isolate types
