+++ src/types/ai.ts
@@ -1,0 +1,1 @@
+export type THmlKey = 'high' | 'mid' | 'low';
