+++ src/types/user.ts
@@ -1,0 +1,5 @@
+export type TUserProps = {
+  nickname: string;
+  email: string;
+  profileUrl: string;
+};
